move the incident serializer to the serializer folder
