+++ src/sentry/rules/actions/notify_event_service.py
@@ -9,7 +9,7 @@ from django import forms
 from sentry.api.serializers import serialize
 from sentry.api.serializers.models.app_platform_event import AppPlatformEvent
 from sentry.eventstore.models import GroupEvent
-from sentry.incidents.endpoints.incident import IncidentSerializer
+from sentry.incidents.endpoints.serializers.incident import IncidentSerializer
 from sentry.incidents.models.alert_rule import AlertRuleTriggerAction
 from sentry.incidents.models.incident import Incident, IncidentStatus
 from sentry.integrations.metric_alerts import incident_attachment_info
